+++ lib/types.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2018. Taimos GmbH http://www.taimos.de
  */
 
-import { RequestEnvelope, ResponseEnvelope } from 'ask-sdk-model';
+import { interfaces, RequestEnvelope, ResponseEnvelope } from 'ask-sdk-model';
 
 export interface SkillSettings {
     /** The skill id */
@@ -24,6 +24,7 @@ export interface InterfaceSettings {
     audio? : boolean;
     video? : boolean;
     apl? : boolean;
+    geolocation? : interfaces.geolocation.GeolocationState;
 }
 
 export interface SequenceItem {