+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix: correct standard card error messages to say "standard card"
 * feat: support all UpsServiceClient calls by adding `distanceUnits`, `temperatureUnits`, and `timeZone` to `withProfile`
 * feat: add `withDeviceAddress` option to test DeviceAddressAPI
+* feat: add Geolocation interface
 
 ## 2.4.0
 * update libraries